[general] misc formatting changes. changed /xdid -R +p to check that atleast 3 points are given.
[button][colorcombo]
misc typecasting changes.

[listview]
changed $xdid().gnum to use a for() loop.
changed /xdid -a to only apply flags as they are needed.
changed /xdid -j to only apply valid listview states.
removed some commented code.

[dcxml]
changed parseCLA() to take a const argument.

[statusbar]
changed /xdid -t to not use ZeroMemory()

[treeview]
removed some duplicate code from /xdid -z
added comment about TreeView_SortChildrenCB() not recursing.
changed insertItem()'s checking of state icon to check for > -1, this fixes the `no state checkbox` problem.
also fixed the same thing in xmlInsertItems()

[xstatusbar]
changed -t to not use ZeroMemory().


git-svn-id: http://dcx.googlecode.com/svn@940 17a06da3-8c20-fa79-550f-e9588e724b77

+++ Classes/dcxbutton.cpp
@@ -704,9 +704,9 @@ void DcxButton::DrawClientArea(HDC hdc, const UINT uMsg, LPARAM lParam)
 
 			if (!this->m_bCtrlCodeText) {
 				if (!this->m_bSelected && this->m_bShadowText)
-					dcxDrawShadowText(hdc,this->m_tsCaption.to_wchr(this->m_bUseUTF8), this->m_tsCaption.wlen(),&rcTxt, DT_WORD_ELLIPSIS | DT_LEFT | DT_TOP | DT_SINGLELINE, this->m_aColors[nState], 0, 5, 5);
+					dcxDrawShadowText(hdc,this->m_tsCaption.to_wchr(this->m_bUseUTF8), (UINT)this->m_tsCaption.wlen(),&rcTxt, DT_WORD_ELLIPSIS | DT_LEFT | DT_TOP | DT_SINGLELINE, this->m_aColors[nState], 0, 5, 5);
 				else
-					DrawTextW( hdc, this->m_tsCaption.to_wchr(this->m_bUseUTF8), this->m_tsCaption.wlen( ), &rcTxt, DT_WORD_ELLIPSIS | DT_LEFT | DT_TOP | DT_SINGLELINE );
+					DrawTextW( hdc, this->m_tsCaption.to_wchr(this->m_bUseUTF8), (int)this->m_tsCaption.wlen( ), &rcTxt, DT_WORD_ELLIPSIS | DT_LEFT | DT_TOP | DT_SINGLELINE );
 			}
 			else
 				mIRC_DrawText(hdc, this->m_tsCaption, &rcTxt, DT_WORD_ELLIPSIS | DT_LEFT | DT_TOP | DT_SINGLELINE, ((!this->m_bSelected && this->m_bShadowText) ? true : false), this->m_bUseUTF8);
